Postavljanje količina namirnica na float #5

+++ Software/app/src/main/java/hr/foi/rampu/fridgium/entities/Namirnica.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Namirnica(
     val id: Int,
     val naziv: String,
-    var kolicina_hladnjak: Int,
+    var kolicina_hladnjak: Float,
     @SerializedName("mjerna_jedinica_id") val mjernaJedinica: MjernaJedinica,
-    val kolicina_kupovina: Int
+    val kolicina_kupovina: Float
 )
